Use simple intrusive linked list for queue
Use a simple intrusive list for the event queue.  The ultimate goal here
is to arrange that certain kinds of events (errors) can be guaranteed
to be queued, as client objects will leak if they are discarded.  As a
secondary improvement, there are some operations (peeking, remove_if)
that can be more efficient as we can hold the queue lock for longer.

This commit is a straight replacement and does not attempt any logic
changes.